requirements.txt
python-telegram-bot
ujson
pytz
django==2.*
sqlparse==0.2.4
djongo
Pillow